ids/requirements.txt
astroid==2.4.2
mysql-connector-python==8.0.21
isort==5.4.2
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
py==1.9.0
pylint==2.6.0
pyshark==0.4.2.11
six==1.15.0
toml==0.10.1
wrapt==1.12.1
